requirements.txt
numpy
Flask
flask-pymongo
bson
librosa
werkzeug
python-dotenv
vosk
soundfile